requirements.txt
streamlit==1.41.1
python-dotenv==1.0.1
tavily-python==0.3.1  # Correct package name
groq==0.3.7
langchain-core==0.2.1
langchain-community==0.2.1
requests==2.32.3
deepseek-chat==0.1.10